+++ src/l1j/jrwz/server/model/L1GroundInventory.java
@@ -48,7 +48,7 @@ public class L1GroundInventory extends L1Inventory {
     private static Logger _log = Logger
             .getLogger(L1PcInventory.class.getName());
 
-    public L1GroundInventory(int objectId, int x, int y, short map) {
+    public L1GroundInventory(int objectId, int x, int y, int map) {
         setId(objectId);
         setX(x);
         setY(y);
